util: small config_option_check.py improvements
As of now this check results in false positives, for instance
'CONFIG_' found in the comments results in an error report.

This patch makes the script a bit more robust:
  - consider only those options mentioned in include/config.h as
    explicitly defined or undefined.
  - do not scan include/config.h for new added CONFIG_ options
  - ignore comments in .mk files

Ideally the script should be scanning only added lines of code and
much smarter about what should be considered a comment and what files
should be examined.

BRANCH=none
BUG=chromium:510672
TEST=the false positives triggered by comments in various build.mk
     files are gone now. Adding an undescribed CONFIG_xxx text still
     triggers the error.

Change-Id: Ib9858775bcd9899dec629682a1c7965e7c2fec96
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285926
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>